修改1.5代码

+++ shardingJdbc1/src/test/java/com/ybwh/springboot2/TesShardingTableByMonth.java
@@ -74,7 +74,7 @@ public class TesShardingTableByMonth {
 //			Integer count = dao.selectCount(dFormat.parse("2018-08-16 16:02:00"), dFormat.parse("2018-09-17 16:27:20"));
 //			System.out.println(count);
 			
-			int sum = dao.selectSum(dFormat.parse("2018-08-17 16:02:00"), dFormat.parse("2018-09-17 16:27:20"));
+			Long sum = dao.selectSum(dFormat.parse("2018-08-17 16:02:00"), dFormat.parse("2018-09-17 16:27:20"));
 			System.out.println(sum);
